[tools] Split update script into launcher and main
